Move repositories into separate namespace
